Moved typeguards to client
